NetworkManager tests: Split WebSocket and libp2p based tests into different files
and add more libp2p tests. It does not seem to be enough yet though, as we currently experience strange disconnects in the web UI which are not reproduces in the tests.

Also rename 'blocklisted' to just blocked. If we have to be politicially correct, let's not overcomplicate things at least :D