CCI: CCI classes now don't enforce cciFamily anymore when fetching Cubes
This is to improve compatibility with CCI-derived classes. On the other hand, appliations are now on their own in making sure that their CubeStore settings and general Cube handling are actually CCI compatible. Maybe we should add some kind of sanity checking and/or CubeFamily coercion back in, but in a less intrusive way.