NetworkPeer: Prevent decreases of lifecycle status
This fixes an issue with the new version of our WebSocket library, which apparently reports the connection ready so incredibly late that we might very well register our high-level handshake as complete before we even register the connection as ready. In this case, the lifecycle status would have stepped back from ONLINE to HANDSHAKING, messing up our tests.