Identity.emitCubeAdded(): Allow Cube as parameter type
This prevents us from creating a CubeInfo before even checking if we need it.

All tests now green; the ones that have been written yet anyway.

Creating CubeInfos in quick succession causes crashes due to uninitialised data, probably due to Github#580. It is curious though that the issue occurs in this context, as we don't actually use Cube's state model but rather create a new Cube object on every store.