Identity.emitCubeAdded(): avoid ping-ponging recursion by keeping track of already visited IDs
While not necessary to prevent infinite recursion (we already prevent that by a simple counter), keeping track of already visited Identities avoids ping-ponging the recursion depth up to the limit, and also prevents duplicate emissions.

It does of course come at the cost of a whole bunch of set operations and may well be more expensive than maxing out the recursion depth counter.