Refactor VeritableBaseImplementation
to incorporate the options object passed on construction rather than picking and choosing individual values.

This is more in line with our general style of keeping and referencing options objects for values which a class will not change internally, and saves us the hassle of demarshalling the options object on construction and potentially remarshalling it later when constructing child objects.

Note that Cube, being the most important user of VeritableBaseImplementation, does not follow the new API (yet). It should, but I don't dare to touch it just now.