cciCockpit: Refactor and make available in every UI controller
Use and incorporate and options object on cciCockpit construction. Make Identity fully optional (i.e. the cockpit also works while the user is logged out)

Require a cciCockpit reference in the cciNode wrapper object as well as in the UI controller context, and make it available as this.cockpit in every UI controller.