Tests: Fix failing tests by relaxing Veritum equality requirements
We recently started retaining certain core fields (DATE, NOTIFY) when restoring Verita from chunks. Those cannot be guaranteed to be restored in the order they were present in the original Veritum, and they currently also do not necessarily demashal in the same order as they are present in the chunk Cubes.