Identity: Drop recursive event emissions and the associated setSubscriptionRecursionDepth() method
Instead, we will use an external RecursiveEmitter for this, which can be obtained using the new getRecursiveEmitter() method.

This:
  - Removes an overly general functionality from our far-too-large Identity class
  - Removes the overly complicated recursion depth state machine build from multiple inter-dependent Identity objects
  - Means that subscribed Identity objects no longer unexpectedly undergo a state (recursion depth) change when the user utilises the recursive emission feature on one Identity

Regression:
  - Already runnning retrievals (be it by event emission or by endless Generator) no longer automatically reflect changes in subscription status, as Identity currently has no way of updating existing RecursiveEmitters

For ZW, this required some ugly hacks to make the current code work; however, the current post retrieval code is still very low-level (i.e. originally based on the CubeStore API) and due to be removed and replaced by high-level Identity-based retrievals.