Cube Subscription end-to-end tests: Refactor to use the Generator provided by CubeRetriever
which is how we expect actual applications to use our API.

Note that this change reveals a timing issue and causes the cancellation test to sometimes fail. This is probably because, by chance, it sometimes runs during a renewal period, and we don't renew subscriptions before they expire yet. I expect this to fix itself once we renew subscriptions ahead of time.