NetworkPeer.handleCubeResponse(): Pass all delivered Cubes to the scheduler at once
There is no longer a need to iterate them one by one on the NetworkPeer layer, that's a relict from when NetworkPeer fed directly to CubeStore

Instead, passing them in a single call allows RequestScheduler to identify them as one delivery, which plays a role in when notification requests are marked fulfilled.

This removes the previous batch processing logic. Rationale: It is no longer needed as we don't spawn a multitude of async requests at a time any more (as we apparently did when feeding directly to CubeStore). Instead, we make a single call to RequestScheduler which then feed them to CubeStore one by one (await after each call). If anything, we should maybe speed that up.