BaseFields: Provide method withoutDisregarded()
which returns a copy of the field set, with all disregarded fields removed.

This will be used to provide a more meaningful equals() comparison.

Disregarded fields are those that appear after a stop field (as defined by the field definition used) and are not positional, as well as the remainder field.
(Note that the stop field itself however is not a disregarded field.)

For example:
  - In the core family, there are no disregarded fields.
    The returned field set will be identical to the original one.
  - In the CCI family, all non-positional fields after CCI_END will be
    disregarded, as will be REMAINDER.