Refactor CubeRetriever.getNotifications()
which will now yield all locally present notifications first, and only then fire a network request.

This is a hack to avoid having to avoid it blocking for a full network timeout in case of unanswered or negative requests.

In the zw microblogging test app, this fixes the issue of subsequent calls to Explore mode lagging.