Identity.getPosts(): Fail gracefully when request options do not match this instance's capability
for example when requesting posts as Verita when this Identity instance has no VeritumRetriever access.

There is a case to be made that we should either stop allowing Identity instances without full capabilities (i.e. always require a VeritumRetriever), but this would result in tighter coupling and more complex test setup (i.e. Identity would become less unit-testable). There is an alternative case to be made that we should throw on unfulfillable options as failing silently can be very confusing, but our general style throughout the Verity lib is currently not to throw.