VeritumRetriever.getNotifications(): Add option to return notifications as Cubes rather than Verita
thus providing a compatible API to CubeRetriever.

This is to ensure that a VeritumRetriever can always be supplied in place of a CubeRetriever, thus making VeritumRetriever a strict augmentation of CubeRetriever.

The broader vision here is to introduce a generalised optional `format` option to each `CubeRetrievalInterface`, supporting retrieval formats such as Cube, CubeInfo or Veritum. Once we have this, we can replace format-specific methods like `getCube`, `getVeritum` and `getCubeInfo` with a single generic `get` method.