Introduce the AUTHORHINT CCI Relationship type
AUTHORHINT allows a Cube to claim a CCI Identity to be this its author. Note that this information is not guaranteed to be true and should never be trusted as such; it is a hint only. Authenticated authorship information is provided by the CCI Identity through MYPOST rels pointing to their content. User agents should only use this hint to check if such a MYPOST rel exists, and disregard the AUTHORHINT otherwise.

The hint is necessary in some use cases to enable the recipient to learn of the (claim) author/Identity. Obviously, our regular way of starting our parsing tree at the Identity root only works if the Identity root Cube is known.