Amend docs
Require the CCI Authorhint relationship to be used no more than once per Veritable and to always be placed in the first Cube. The latter is to enable clients to resolve (and verify) the authorship claim before downloading the whole of a potentially very long Veritum.

Identity: Amend comments to mark a TODO