NetworkPeer: Make indirect notification requests work (Github PR#876)
This is a **breaking** change to our wire format. (It does however only break indirect notification requests which never worked to begin with, so maybe breaking is not as bad as it sounds here.)

This rewrites the network protocol for KeyRequests in Notifications-by-Difficulty and Notifications-by-Timestamp mode. The main change here is that requesting nodes are now required to provide a lower bound database key for the appropriate mode, which is a concatenation of the notification key (recipient), either minimum difficulty or minimum timestamp, and the notification's Cube key.

Node that Notification-by-Timestamps are currently unused. The networking core supports them and we have an index sublevel for them, but nobody ever actually fires requests in that mode.

Squashed commits:

* NetworkPeer tests: Add basic tests regarding indirect notification requests

Lo and behold, they fail. Did that crap never actually work?! :D

* LevelBackend: Provide *inefficient and thus deprecated* method autocompletePartialKey() to retrieve full database keys based on partials

CubeStore.getKeyRange(): Offer auto-conversion of CubeKeys to database keys when operating on sublevels other than CUBES

* networkPeer.notificationRequest.test.ts: Verity test setup

* CubeStore tests: Fix key range iteration tests for notification DBs

We only have tests using explicit database keys

* NetworkPeer: Rewrite notification key requests

Previous implementation was completely broken

* CubeStore CubeIteratorOptionsSublevel: Remove option autoConvertInputKeys

as it is not needed for core operations and there is no efficient implementation for it (best we can do is O(n))

* NetworkPeer notification request tests: Fix faulty test

* Fix typo