Revert "chunkDecryption.ts: throw if veritum decryption fails for easier debugging (#836)"
This reverts commit 89c0b3e7fa7e27adc138be8b8becd3e372f9b3ce.

We don't throw towards application code; breaking that contract randomly
breaks applications downstream. There's a choice between returning
undefined on failure or throwing; both have advantages and disadvantages. Using
both at the same time does not add any value. It's just as easy to check
for undefined in tests as it is to expect an exception.